changes to support dao/master test script design AND NOTES since i dont know when I will be able to touch this again
1. added dao and unit test
to acquire information from a master test script
2. added preliminary command/receiver/invoker patter to take the
commands from the master test script and propogate it out to the
supporting objects
3. started to modify the database interfaces to allow acquisition from
other databases other than an excel spreadsheet

NOTE:  its been awhile since I checked anything else in and leaving this
as note.  overall direction is to get the dao working and set.  right
now, the dao only works for the master test script.   I would like to
make this a bit more generic so that I can apply it to other tables in
the spreadsheet.  FOCUS should be on treating them as tables so that
they can move between datasources.   the other thing is the commander. 
right now the interfaces do not have any required methods so this is in
early design.   there are two testcases in the test package. use these
tests to prove you can move between any of the runners as more changes
are made.